Seeder

+++ database/migrations/2020_07_03_131552_create_houses_table.php
@@ -21,7 +21,7 @@ class CreateHousesTable extends Migration
             $table->tinyInteger('rooms');
             $table->tinyInteger('beds');
             $table->tinyInteger('bathrooms');
-            $table->tinyInteger('sqm');
+            $table->smallInteger('sqm');
             $table->string('address');
             $table->string('lat');
             $table->string('long');
@@ -30,7 +30,7 @@ class CreateHousesTable extends Migration
             $table->timestamps();
         });
     }
-    
+
 
     /**
      * Reverse the migrations.
